fix: firestore.tsからinitialize部分をfirebase.tsに移行
